# backend/server.js
import express from 'express';
import jwt from 'jsonwebtoken';
import bcrypt from 'bcrypt';
import morgan from 'morgan';
import { PGlite } from "@electric-sql/pglite";
import { v4 as uuidv4 } from 'uuid';
import dotenv from 'dotenv';

// Initializing environment variables
dotenv.config();

const app = express();
app.use(express.json());
app.use(morgan('dev'));

// Enable CORS
app.use((req, res, next) => {
  res.header('Access-Control-Allow-Origin', '*');
  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
  res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
  next();
});

const PORT = process.env.PORT || 1337;
const DB = './db';
const postgres = new PGlite(DB);

// JWT Middleware for protected routes
const authenticateJWT = (req, res, next) => {
  const authHeader = req.headers.authorization;
  if (authHeader) {
    const token = authHeader.split(' ')[1];
    jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
      if (err) {
        return res.sendStatus(403);
      }
      req.user = user;
      next();
    });
  } else {
    res.sendStatus(401);
  }
};

// Utility function to generate JWT
const generateAccessToken = (userId) => {
  return jwt.sign({ userId }, process.env.JWT_SECRET, { expiresIn: '24h' });
};

// Utility function to hash password
const hashPassword = async (password) => {
  return await bcrypt.hash(password, 10);
};

// 1. GET /api/ai_tools - Search and Filter AI Tools
app.get('/api/ai_tools', async (req, res) => {
  try {
    const { pricing, tags, name } = req.query;
    let query = "SELECT * FROM ai_tools WHERE 1=1";
    const params = [];

    if (pricing) {
      query += " AND pricing = ?";
      params.push(pricing);
    }
    if (tags) {
      query += " AND tags ILIKE ?";
      params.push(`%${tags}%`);
    }
    if (name) {
      query += " AND name ILIKE ?";
      params.push(`%${name}%`);
    }

    query += " LIMIT 5";  // Limiting results

    const results = await postgres.query(query, params);
    res.status(200).json({ tools: results.rows });
  } catch (err) {
    res.status(500).json({ error: "Error fetching AI Tools directory" });
  }
});

// 2. POST /api/users - Register new user
app.post('/api/users', async (req, res) => {
  try {
    const { email, password, first_name, last_name } = req.body;
    const user_id = uuidv4();

    // Hash the password
    const password_hash = await hashPassword(password);
    const created_at = new Date().toISOString();

    const insertQuery = `
      INSERT INTO users (user_id, email, password_hash, first_name, last_name, created_at) 
      VALUES ($1, $2, $3, $4, $5, $6)
    `;
    await postgres.query(insertQuery, [user_id, email, password_hash, first_name, last_name, created_at]);

    const token = generateAccessToken(user_id);

    res.status(201).json({ token });
  } catch (err) {
    res.status(500).json({ error: "User creation failed" });
  }
});

// 3. POST /api/users/login - Login user
app.post('/api/users/login', async (req, res) => {
  const { email, password } = req.body;

  try {
    const result = await postgres.query("SELECT * FROM users WHERE email = $1", [email]);

    if (result.rows.length === 0) {
      return res.status(401).json({ error: 'Invalid credentials' });
    }

    const user = result.rows[0];
    const isValidPassword = await bcrypt.compare(password, user.password_hash);

    if (isValidPassword) {
      const token = generateAccessToken(user.user_id);
      return res.status(200).json({ token });
    } else {
      return res.status(401).json({ error: 'Invalid credentials' });
    }
  } catch (err) {
    res.status(500).json({ error: "Login failed" });
  }
});

// 4. POST /api/workflows - Create a new custom workflow
app.post('/api/workflows', authenticateJWT, async (req, res) => {
  const { title, category, description, nodes } = req.body;
  const user_id = req.user.userId;
  const workflow_id = uuidv4();
  const created_at = new Date().toISOString();

  try {
    await postgres.query(
      "INSERT INTO workflows (workflow_id, user_id, title, category, description, is_predefined, created_at) VALUES ($1, $2, $3, $4, $5, false, $6)",
      [workflow_id, user_id, title, category, description, created_at]
    );

    const nodePromises = nodes.map(node => {
      const node_id = uuidv4();
      const { title, description } = node;
      return postgres.query(
        "INSERT INTO workflow_nodes (node_id, workflow_id, title, description, created_at) VALUES ($1, $2, $3, $4, $5)",
        [node_id, workflow_id, title, description, created_at]
      );
    });

    await Promise.all(nodePromises);

    res.status(201).json({ workflow_id, title, category, description, user_id });
  } catch (err) {
    res.status(500).json({ error: "Error creating workflow" });
  }
});

// 5. GET /api/workflows - Retrieve all user-created workflows
app.get('/api/workflows', authenticateJWT, async (req, res) => {
  try {
    const user_id = req.user.userId;
    const results = await postgres.query("SELECT * FROM workflows WHERE user_id = $1", [user_id]);
    res.status(200).json(results.rows);
  } catch (err) {
    res.status(500).json({ error: "Error fetching user's workflows" });
  }
});

// 6. GET /api/workflows/:workflow_id - Get specific workflow details
app.get('/api/workflows/:workflow_id', authenticateJWT, async (req, res) => {
  try {
    const { workflow_id } = req.params;
    const workflowQuery = "SELECT * FROM workflows WHERE workflow_id = $1";
    const workflowResults = await postgres.query(workflowQuery, [workflow_id]);

    if (workflowResults.rows.length === 0) {
      return res.status(404).json({ error: "Workflow not found" });
    }

    res.status(200).json(workflowResults.rows[0]);
  } catch (err) {
    res.status(500).json({ error: "Error fetching workflow details" });
  }
});

// 7. GET /api/workflows/:workflow_id/nodes/:node_id/suggestions - Show AI tool suggestions for a workflow node
app.get('/api/workflows/:workflow_id/nodes/:node_id/suggestions', authenticateJWT, async (req, res) => {
  try {
    const { node_id } = req.params;
    const query = `
      SELECT ai_tools.* 
      FROM ai_tools
      INNER JOIN ai_tool_node_suggestions ON ai_tools.tool_id = ai_tool_node_suggestions.tool_id
      WHERE ai_tool_node_suggestions.node_id = $1
      ORDER BY ai_tool_node_suggestions.rank ASC
      LIMIT 5
    `;
    const results = await postgres.query(query, [node_id]);

    res.status(200).json({ suggestions: results.rows });
  } catch (err) {
    res.status(500).json({ error: "Error fetching AI tool suggestions" });
  }
});

// 8. PUT /api/workflows/:workflow_id - Edit a workflow
app.put('/api/workflows/:workflow_id', authenticateJWT, async (req, res) => {
  const { workflow_id } = req.params;
  const { title, category, description, nodes } = req.body;

  try {
    await postgres.query(
      "UPDATE workflows SET title = $1, category = $2, description = $3 WHERE workflow_id = $4",
      [title, category, description, workflow_id]
    );

    // Optionally handle updating nodes (nodes can be complex, so skipping for simplicity)

    res.status(200).json({ workflow_id, title, category, description });
  } catch (err) {
    res.status(500).json({ error: "Error updating workflow" });
  }
});

app.listen(PORT, () => console.log(`Server is running on http://localhost:${PORT}`));